Move all interfaces to interface.ts for refactoring
